Implemented NFO Export

+++ resources/lib/kodi/library.py
@@ -4,6 +4,8 @@ from __future__ import unicode_literals
 
 import os
 import re
+import sys
+
 from datetime import datetime, timedelta
 from functools import wraps
 
@@ -14,6 +16,10 @@ from resources.lib.globals import g
 import resources.lib.common as common
 import resources.lib.api.shakti as api
 import resources.lib.kodi.ui as ui
+import resources.lib.kodi.nfo as nfo
+
+import xml.etree.ElementTree as ET
+
 
 LIBRARY_HOME = 'library'
 FOLDER_MOVIES = 'movies'
@@ -89,9 +95,10 @@ def _get_item(mediatype, filename):
             {'field': 'path', 'operator': 'startswith', 'value': path},
             {'field': 'filename', 'operator': 'is', 'value': shortname}
             ]})[0]
+    if not library_item:
+        raise ItemNotFound
     return common.get_library_item_details(
          mediatype, library_item[mediatype + 'id'])
-    raise ItemNotFound
 
 
 def list_contents():
@@ -116,7 +123,7 @@ def update_kodi_library(library_operation):
             _remove_from_kodi_library(videoid)
         library_operation(videoid, task_handler, *args, **kwargs)
         if not is_remove:
-            # Update kodi library through service
+            # Update Kodi library through service
             # This prevents a second call to cancel the update
             common.debug('Notify service to update the library')
             common.send_signal(common.Signals.LIBRARY_UPDATE_REQUESTED)
@@ -129,7 +136,7 @@ def _remove_from_kodi_library(videoid):
     try:
         kodi_library_items = [get_item(videoid)]
         if videoid.mediatype == common.VideoId.SHOW or videoid.mediatype == common.VideoId.SEASON:
-            # Retreive the all episodes in the export folder
+            # Retrieve the all episodes in the export folder
             filters = {'and': [
                 {'field': 'path', 'operator': 'startswith', 'value': os.path.dirname(kodi_library_items[0]['file'])},
                 {'field': 'filename', 'operator': 'endswith', 'value': '.strm'}
@@ -174,70 +181,89 @@ def purge():
 
 
 @common.time_execution(immediate=False)
-def compile_tasks(videoid):
+def compile_tasks(videoid, nfo_settings=None):
     """Compile a list of tasks for items based on the videoid"""
     common.debug('Compiling library tasks for {}'.format(videoid))
     metadata = api.metadata(videoid)
     if videoid.mediatype == common.VideoId.MOVIE:
-        return _create_movie_task(videoid, metadata[0])
+        return _create_movie_task(videoid, metadata[0], nfo_settings)
     elif videoid.mediatype in common.VideoId.TV_TYPES:
-        return _create_tv_tasks(videoid, metadata)
-
+        return _create_tv_tasks(videoid, metadata, nfo_settings)
     raise ValueError('Cannot handle {}'.format(videoid))
 
 
-def _create_movie_task(videoid, movie):
+def _create_movie_task(videoid, movie, nfo_settings):
     """Create a task for a movie"""
+    # Reset NFO export to false if we never want movies nfo
     name = '{title} ({year})'.format(title=movie['title'], year=movie['year'])
-    return [_create_item_task(name, FOLDER_MOVIES, videoid, name, name)]
+    return [_create_item_task(name, FOLDER_MOVIES, videoid, name, name,
+                              nfo.create_movie_nfo(movie) if
+                              nfo_settings and nfo_settings.export_movie_enabled else None)]
 
 
-def _create_tv_tasks(videoid, metadata):
+def _create_tv_tasks(videoid, metadata, nfo_settings):
     """Create tasks for a show, season or episode.
     If videoid represents a show or season, tasks will be generated for
     all contained seasons and episodes"""
     if videoid.mediatype == common.VideoId.SHOW:
-        return _compile_show_tasks(videoid, metadata[0])
+        tasks = _compile_show_tasks(videoid, metadata[0], nfo_settings)
     elif videoid.mediatype == common.VideoId.SEASON:
-        return _compile_season_tasks(videoid, metadata[0],
-                                     common.find(int(videoid.seasonid), 'id',
-                                                 metadata[0]['seasons']))
-    return [_create_episode_task(videoid, *metadata)]
+       tasks = _compile_season_tasks(videoid,
+                                     metadata[0],
+                                     common.find(int(videoid.seasonid),
+                                                 'id',
+                                                 metadata[0]['seasons']),
+                                     nfo_settings)
+    else:
+        tasks = [_create_episode_task(videoid, *metadata, nfo_settings=nfo_settings)]
+
+    if nfo_settings and nfo_settings.export_full_tvshow:
+        # Create tvshow.nfo file
+        tasks.append(_create_item_task('tvshow.nfo', FOLDER_TV, videoid,
+                                       metadata[0]['title'],
+                                       'tvshow',
+                                       nfo.create_show_nfo(metadata[0]),
+                                       False))
+    return tasks
 
 
-def _compile_show_tasks(videoid, show):
+def _compile_show_tasks(videoid, show, nfo_settings):
     """Compile a list of task items for all episodes of all seasons
     of a tvshow"""
-    # This nested comprehension is nasty but neccessary. It flattens
+    # This nested comprehension is nasty but necessary. It flattens
     # the task lists for each season into one list
     return [task for season in show['seasons']
             for task in _compile_season_tasks(
-                videoid.derive_season(season['id']), show, season)]
+                videoid.derive_season(season['id']), show, season, nfo_settings)]
 
 
-def _compile_season_tasks(videoid, show, season):
+def _compile_season_tasks(videoid, show, season, nfo_settings):
     """Compile a list of task items for all episodes in a season"""
     return [_create_episode_task(videoid.derive_episode(episode['id']),
-                                 episode, season, show)
+                                 episode, season, show, nfo_settings)
             for episode in season['episodes']]
 
 
-def _create_episode_task(videoid, episode, season, show):
+def _create_episode_task(videoid, episode, season, show, nfo_settings):
     """Export a single episode to the library"""
     filename = 'S{:02d}E{:02d}'.format(season['seq'], episode['seq'])
     title = ' - '.join((show['title'], filename, episode['title']))
     return _create_item_task(title, FOLDER_TV, videoid, show['title'],
-                             filename)
+                             filename,
+                             nfo.create_episode_nfo(episode, season, show)
+                             if nfo_settings and nfo_settings.export_tvshow_enabled else None)
 
 
-def _create_item_task(title, section, videoid, destination, filename):
+def _create_item_task(title, section, videoid, destination, filename, nfo_data=None, is_strm=True):
     """Create a single task item"""
     return {
         'title': title,
         'section': section,
         'videoid': videoid,
         'destination': re.sub(ILLEGAL_CHARACTERS, '', destination),
-        'filename': re.sub(ILLEGAL_CHARACTERS, '', filename)
+        'filename': re.sub(ILLEGAL_CHARACTERS, '', filename),
+        'nfo_data': nfo_data,
+        'is_strm': is_strm
     }
 
 
@@ -247,11 +273,16 @@ def export_item(item_task, library_home):
     # Paths must be legal to ensure NFS compatibility
     destination_folder = xbmc.makeLegalFilename(os.path.join(
         library_home, item_task['section'], item_task['destination']))
-    export_filename = xbmc.makeLegalFilename(os.path.join(
-        destination_folder.decode('utf-8'), item_task['filename'] + '.strm'))
-    _add_to_library(item_task['videoid'], export_filename)
     _create_destination_folder(destination_folder)
-    _write_strm_file(item_task, export_filename)
+    if item_task['is_strm']:
+        export_filename = xbmc.makeLegalFilename(os.path.join(
+            destination_folder.decode('utf-8'), item_task['filename'] + '.strm'))
+        _add_to_library(item_task['videoid'], export_filename)
+        _write_strm_file(item_task, export_filename)
+    if item_task['nfo_data'] is not None:
+        nfo_filename = xbmc.makeLegalFilename(os.path.join(
+            destination_folder.decode('utf-8'), item_task['filename'] + '.nfo'))
+        _write_nfo_file(item_task['nfo_data'], nfo_filename)
     common.debug('Exported {}'.format(item_task['title']))
 
 
@@ -268,6 +299,16 @@ def _write_strm_file(item_task, export_filename):
     try:
         filehandle.write(common.build_url(videoid=item_task['videoid'],
                                           mode=g.MODE_PLAY).encode('utf-8'))
+    finally:
+        filehandle.close()
+
+
+def _write_nfo_file(nfo_data, nfo_filename):
+    """Write the NFO file"""
+    filehandle = xbmcvfs.File(xbmc.translatePath(nfo_filename), 'w')
+    try:
+        filehandle.write('<?xml version=\'1.0\' encoding=\'UTF-8\'?>'.encode('utf-8'))
+        filehandle.write(ET.tostring(nfo_data, encoding='utf-8', method='xml'))
     finally:
         filehandle.close()
 
@@ -291,26 +332,39 @@ def _add_to_library(videoid, export_filename):
 def remove_item(item_task, library_home=None):
     """Remove an item from the library and delete if from disk"""
     # pylint: disable=unused-argument, broad-except
-    common.debug('Removing {} from library'.format(item_task['title']))
-    if not is_in_library(item_task['videoid']):
-        common.warn('cannot remove {}, item not in library'
-                    .format(item_task['title']))
-        return
-    id_path = item_task['videoid'].to_list()
-    exported_filename = xbmc.translatePath(
-        common.get_path(id_path, g.library())['file']).decode("utf-8")
-    parent_folder = os.path.dirname(exported_filename)
-    try:
-        xbmcvfs.delete(xbmc.translatePath(exported_filename).decode("utf-8"))
-        # Fix parent folder not removed
-        dirs, files = xbmcvfs.listdir(xbmc.translatePath(parent_folder).decode("utf-8"))
-        if not dirs and not files: # the destination folder is empty
-            xbmcvfs.rmdir(xbmc.translatePath(parent_folder).decode("utf-8"))
-    except Exception:
-        common.debug('Cannot delete {}, file does not exist'
-                     .format(exported_filename))
-    common.remove_path(id_path, g.library(), lambda e: e.keys() == ['videoid'])
-    g.save_library()
+    if item_task['is_strm']:  # We don't take care of a tvshow.nfo task if we are running an update
+        common.debug('Removing {} from library'.format(item_task['title']))
+        if not is_in_library(item_task['videoid']):
+            common.warn('cannot remove {}, item not in library'
+                        .format(item_task['title']))
+            return
+        id_path = item_task['videoid'].to_list()
+        exported_filename = xbmc.translatePath(
+            common.get_path(id_path, g.library())['file']).decode("utf-8")
+        parent_folder = os.path.dirname(exported_filename)
+        try:
+            xbmcvfs.delete(xbmc.translatePath(exported_filename).decode("utf-8"))
+            # Remove the NFO files if exists
+            nfo_file = os.path.splitext(xbmc.translatePath(exported_filename).decode("utf-8"))[0]+'.nfo'
+            if xbmcvfs.exists(nfo_file):
+                xbmcvfs.delete(nfo_file)
+            dirs, files = xbmcvfs.listdir(xbmc.translatePath(parent_folder).decode("utf-8"))
+            tvshow_nfo_file = xbmc.makeLegalFilename(
+                os.path.join(
+                    xbmc.translatePath(parent_folder).decode("utf-8"), 'tvshow.nfo'))
+            # Remove tvshow_nfo_file only when is the last file (users have the option of removing even single seasons)
+            if xbmcvfs.exists(tvshow_nfo_file) and not dirs and len(files) == 1:
+                xbmcvfs.delete(tvshow_nfo_file)
+                # Delete parent folder
+                xbmcvfs.rmdir(xbmc.translatePath(parent_folder).decode("utf-8"))
+            # Delete parent folder when empty
+            if not dirs and not files:
+                xbmcvfs.rmdir(xbmc.translatePath(parent_folder).decode("utf-8"))
+        except Exception:
+            common.debug('Cannot delete {}, file does not exist'
+                         .format(exported_filename))
+        common.remove_path(id_path, g.library(), lambda e: e.keys() == ['videoid'])
+        g.save_library()
 
 
 def update_item(item_task, library_home):
@@ -345,10 +399,10 @@ def update_library():
 
 
 @update_kodi_library
-def execute_library_tasks(videoid, task_handler, title, sync_mylist=True):
+def execute_library_tasks(videoid, task_handler, title, sync_mylist=True, nfo_settings=None):
     """Execute library tasks for videoid and show errors in foreground"""
     common.execute_tasks(title=title,
-                         tasks=compile_tasks(videoid),
+                         tasks=compile_tasks(videoid, nfo_settings),
                          task_handler=task_handler,
                          notify_errors=True,
                          library_home=library_path())
@@ -356,10 +410,10 @@ def execute_library_tasks(videoid, task_handler, title, sync_mylist=True):
 
 
 @update_kodi_library
-def execute_library_tasks_silently(videoid, task_handler, sync_mylist):
+def execute_library_tasks_silently(videoid, task_handler, sync_mylist, nfo_settings=None):
     """Execute library tasks for videoid and don't show any GUI feedback"""
     # pylint: disable=broad-except
-    for task in compile_tasks(videoid):
+    for task in compile_tasks(videoid, nfo_settings):
         try:
             task_handler(task, library_path())
         except Exception:
